Fix: receive hang indefinitely
After writing the first bytes from buffer, it doesn't "notify" the reader
and continue to write to the buffer without knowing buffer is consume entirely or not.

The reader hence can send wrong data to receiver. With incorrect box size, it can hang forever.

`testpic` segment size is small (<65536), so it is not affected.